deployment/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
openai
python-dotenv
python-multipart
fpdf2
werkzeug
gunicorn
asgiref 